create login route
also test routes and modularise server and routes. relates #5
