+++ config.env
@@ -1,1 +1,3 @@
-BASE_URL = https://localhost:3000/
+BASE_URL=https://localhost:3000/
+export CLIENT_ID=16fb48ff6848f1be683e
+export CLIENT_SECRET=69b012dad1dbbb4c7ff194dc0b8c37e0ed450460
