+++ package.json
@@ -4,8 +4,8 @@
   "description": "week-8-oauth-alice-lucy",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node src/server.js"
+    "test": "tape tests/test.js | tap-spec",
+    "start": "node src/start.js"
   },
   "repository": {
     "type": "git",
@@ -18,8 +18,14 @@
   },
   "homepage": "https://github.com/FAC10/week-8-oauth-alice-lucy#readme",
   "dependencies": {
+    "env2": "^2.1.1",
     "hapi": "^16.1.1",
     "hapi-require-https": "^2.1.0",
-    "inert": "^4.2.0"
+    "inert": "^4.2.0",
+    "shot": "^3.4.0"
+  },
+  "devDependencies": {
+    "tap-spec": "^4.1.1",
+    "tape": "^4.6.3"
   }
 }
