+++ src/routes/home.js
@@ -1,0 +1,7 @@
+module.exports = {
+  method: 'GET',
+  path: '/',
+  handler: function (request, reply) {
+    reply.file('public/index.html');
+  }
+}
