+++ src/routes/index.js
@@ -1,0 +1,4 @@
+module.exports = [
+  require('./home.js'),
+  require('./login.js')
+]
