+++ src/server.js
@@ -5,8 +5,8 @@ const Hapi = require('hapi');
 const server = new Hapi.Server();
 const Path = require('path');
 const Inert = require('inert');
-const HTTPS = require('hapi-require-https');
 const fs = require('fs');
+const Routes = require('./routes');
 //
 
 server.connection({port: 3000, host: 'localhost',
@@ -20,18 +20,7 @@ server.register(Inert, (err) => {
   if (err) {
     throw err;
   }
-  server.route({
-    method: 'GET',
-    path: '/',
-    handler: function (request, reply) {
-      reply.file('public/index.html');
-    }
-  });
+  server.route(Routes);
 });
 
-server.start((err) => {
-  if (err) {
-    throw err;
-  }
-  console.log(`Server running at: ${server.info.uri}`);
-});
+module.exports = server;
