+++ src/start.js
@@ -1,8 +1,8 @@
-// module.exports = start
-//
-// start = server.start((err) {
-//   if (err) {
-//     throw err;
-//   }
-//   console.log(`Server running at: ${port}`);
-// });
+const server = require('./server.js');
+
+server.start((err) => {
+  if (err) {
+    throw err;
+  }
+  console.log(`Server running at: ${server.info.uri}`);
+});
